wip: Ajeitando arquivos e componentização. #HEW-13
