+++ frontend/src/components/footer.jsx
@@ -1,0 +1,11 @@
+import '../App.css';
+
+function Footer() {
+    return (
+      <div>
+        <h1>footer</h1>
+      </div>
+    );
+  }
+  
+export default Footer;
