+++ frontend/src/main.jsx
@@ -1,11 +1,13 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import PaginaInicial from './pages/pagina-inicial.jsx'
-
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
+import { BrowserRouter as Router } from 'react-router-dom';
+import './index.css';
+import PaginaInicial from './pages/PaginaInicial/pagina-inicial.jsx';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <PaginaInicial />
-  </StrictMode>,
-)
+    <Router>
+      <PaginaInicial />
+    </Router>
+  </StrictMode>
+);
