+++ frontend/src/pages/Cadastro/cadastro.jsx
@@ -1,0 +1,11 @@
+import '../../App.css';
+
+function Cadastro() {
+  return (
+    <div className="cadastro-container">
+      aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
+    </div>
+  );
+}
+
+export default Cadastro;
