+++ frontend/src/pages/PaginaInicial/hemocentros.jsx
@@ -1,0 +1,11 @@
+import '../../App.css';
+
+function Hemocentros() {
+    return (
+      <div className="hemocentros-container">
+        hemocentros
+      </div>
+    );
+  }
+
+export default Hemocentros;
