+++ frontend/src/pages/PaginaInicial/message.jsx
@@ -1,0 +1,11 @@
+import '../../App.css';
+
+function Message() {
+    return (
+      <div className="message-container">
+        mensagem
+      </div>
+    );
+  }
+
+export default Message;
